+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fixed: InMemory search failing to query correctly when just the table name is provided
 * Fixed: InMemory concurrency issues with extended data simultaneous deletes and adds
 * Fixed: getExtendedData single row with multiple matching row id prefixes
+* Fixed: Elasticsearch limit while using scroll API 
 
 # v4.8.1
 * Added: Term aggregation result includes `sum_other_doc_count` and `doc_count_error_upper_bound`